refactor(lists): split Lists dashboard into index + show pages
The Lists dashboard crammed the whole feature onto one ~1475-line page
(left rail of lists + right-pane editor) and creating a list bounced you
back to the collection instead of the list you just made. Split it into a
collection page and a per-list detail page, mirroring the overlays UX.

- index.vue: slim collection view with a client-side search box that
  filters by slug, label, AND item contents (content-only hits show a
  "matches:" hint), compact rows linking to show, the create modal, and
  the global !list meta-command panel. Echo subscription kept for live rows.
- show.vue (new): per-list page with items editor, disable toggle, expiry,
  actions + chat-permissions, snapshots, and append commands.
- ListController: new show($slug) resolved scoped to the user (per-user
  slugs make implicit {list:slug} binding unsafe); store() redirects to
  lists.show, destroy() redirects to lists.index.
- routes/web.php: GET /{slug} named lists.show, after the meta-command
  literals so it can't swallow them.
- store test asserts the new redirect target. 98 list tests pass.
